+++ deepcave/plugins/objective/pareto_front.py
@@ -17,9 +17,11 @@ import numpy as np
 import plotly.graph_objs as go
 from dash import dcc, html
 
+from deepcave import notification
 from deepcave.config import Config
 from deepcave.plugins.dynamic import DynamicPlugin
 from deepcave.runs import AbstractRun, Status, check_equality
+from deepcave.runs.exceptions import NotMergeableError, RunInequality
 from deepcave.utils.layout import get_select_options, help_button
 from deepcave.utils.styled_plot import plt
 from deepcave.utils.styled_plotty import (
@@ -55,6 +57,9 @@ class ParetoFront(DynamicPlugin):
         Since this function is called before the layout is created,
         it can be also used to set common values for the plugin.
 
+        If the runs are not mergeable, they still should be displayed
+        but with a corresponding warning message
+
         Parameters
         ----------
         runs : List[AbstractRun]
@@ -68,18 +73,36 @@ class ParetoFront(DynamicPlugin):
             If the budgets of the runs are not equal.
             If the objective of the runs are not equal.
         """
-        check_equality(runs, objectives=True, budgets=True)
+        try:
+            check_equality(runs, objectives=True, budgets=True)
+        except NotMergeableError as e:
+            run_inequality = e.args[1]
+            if run_inequality == RunInequality.INEQ_BUDGET:
+                notification.update("The budgets of the runs are not equal.", color="warning")
+            elif run_inequality == RunInequality.INEQ_CONFIGSPACE:
+                notification.update(
+                    "The configuration spaces of the runs are not equal.", color="warning"
+                )
+            elif run_inequality == RunInequality.INEQ_META:
+                notification.update("The meta data of the runs is not equal.", color="warning")
+            elif run_inequality == RunInequality.INEQ_OBJECTIVE:
+                raise NotMergeableError("The objectives of the selected runs cannot be merged.")
 
         # Set some attributes here
-        run = runs[0]
+        # It is necessary to get the run with the smallest budget and objective options
+        # as first comparative value, else there is gonna be an index problem
+        objective_options = []
+        budget_options = []
+        for run in runs:
+            objective_names = run.get_objective_names()
+            objective_ids = run.get_objective_ids()
+            objective_options.append(get_select_options(objective_names, objective_ids))
 
-        objective_names = run.get_objective_names()
-        objective_ids = run.get_objective_ids()
-        self.objective_options = get_select_options(objective_names, objective_ids)
-
-        budgets = run.get_budgets(human=True)
-        budget_ids = run.get_budget_ids()
-        self.budget_options = get_select_options(budgets, budget_ids)
+            budgets = run.get_budgets(human=True)
+            budget_ids = run.get_budget_ids()
+            budget_options.append(get_select_options(budgets, budget_ids))
+        self.objective_options = min(objective_options, key=len)
+        self.budget_options = min(budget_options, key=len)
 
     @staticmethod
     def get_input_layout(register: Callable) -> List[Any]: